+++ azext_iot/sdk/pnp/digital_twin_repository_service.py
@@ -15,7 +15,7 @@ from .version import VERSION
 from msrest.pipeline import ClientRawResponse
 from msrest.exceptions import HttpOperationError
 from . import models
-from azext_iot.constants import VERSION as extver
+from azext_iot.constants import USER_AGENT
 
 
 class DigitalTwinRepositoryServiceConfiguration(Configuration):
@@ -34,7 +34,7 @@ class DigitalTwinRepositoryServiceConfiguration(Configuration):
 
         super(DigitalTwinRepositoryServiceConfiguration, self).__init__(base_url)
         self.add_user_agent('digitaltwinrepositoryservice/{}'.format(VERSION))
-        self.add_user_agent('IoTPlatformCliExtension/{}'.format(extver))
+        self.add_user_agent(USER_AGENT)
 
 
 class DigitalTwinRepositoryService(SDKClient):